+++ src/Ares/Rule/BlankableRule.php
@@ -13,14 +13,25 @@ namespace Ares\Rule;
 
 use Ares\Context;
 use Ares\Exception\InvalidValidationRuleArgsException;
+use Ares\Schema\Type;
 
 /**
  * Class BlankableRule
  */
-class BlankableRule implements RuleInterface
+class BlankableRule extends AbstractRule
 {
-    const ID            = 'blankable';
-    const ERROR_MESSAGE = 'Value must not be blank';
+    public const ID            = 'blankable';
+    public const ERROR_MESSAGE = 'Value must not be blank';
+
+    /**
+     * @return array
+     */
+    public function getSupportedTypes(): array
+    {
+        return [
+            Type::STRING,
+        ];
+    }
 
     /**
      * @param mixed         $args    Validation rule configuration.
@@ -29,7 +40,7 @@ class BlankableRule implements RuleInterface
      * @return boolean
      * @throws \Ares\Exception\InvalidValidationRuleArgsException
      */
-    public function validate($args, $data, Context $context): bool
+    public function performValidation($args, $data, Context $context): bool
     {
         if (!is_bool($args)) {
             throw new InvalidValidationRuleArgsException('Invalid args: ' . json_encode($args));