+++ src/Ares/Rule/MinRule.php
@@ -12,7 +12,6 @@ declare(strict_types=1);
 namespace Ares\Rule;
 
 use Ares\Context;
-use Ares\Exception\InapplicableValidationRuleException;
 use Ares\Exception\InvalidValidationRuleArgsException;
 use Ares\Rule\TypeRule;
 use Ares\Schema\Type;
@@ -20,27 +19,32 @@ use Ares\Schema\Type;
 /**
  * Class MinRule
  */
-class MinRule implements RuleInterface
+class MinRule extends AbstractRule
 {
-    const ID            = 'min';
-    const ERROR_MESSAGE = 'Value must not be smaller than {value}';
+    public const ID            = 'min';
+    public const ERROR_MESSAGE = 'Value must not be smaller than {value}';
+
+    /**
+     * @return array
+     */
+    public function getSupportedTypes(): array
+    {
+        return [
+            Type::FLOAT,
+            Type::INTEGER,
+            Type::NUMERIC,
+        ];
+    }
 
     /**
      * @param mixed         $args    Validation rule configuration.
      * @param mixed         $data    Input data.
      * @param \Ares\Context $context Validation context.
      * @return boolean
-     * @throws \Ares\Exception\InapplicableValidationRuleException
      * @throws \Ares\Exception\InvalidValidationRuleArgsException
      */
-    public function validate($args, $data, Context $context): bool
+    public function performValidation($args, $data, Context $context): bool
     {
-        $schema = $context->getSchema();
-
-        if (!in_array($schema->getRule(TypeRule::ID)->getArgs(), [Type::FLOAT, Type::INTEGER], true)) {
-            throw new InapplicableValidationRuleException('This rule applies to <float> and <integer> types only');
-        }
-
         if (!is_numeric($args)) {
             throw new InvalidValidationRuleArgsException('Invalid args: ' . json_encode($args));
         }