+++ src/Ares/RuleFactory.php
@@ -15,8 +15,11 @@ use Ares\Exception\UnknownValidationRuleIdException;
 use Ares\Rule\AllowedRule;
 use Ares\Rule\BlankableRule;
 use Ares\Rule\DateTimeRule;
+use Ares\Rule\DirectoryRule;
 use Ares\Rule\EmailRule;
+use Ares\Rule\FileRule;
 use Ares\Rule\ForbiddenRule;
+use Ares\Rule\LengthRule;
 use Ares\Rule\MaxLengthRule;
 use Ares\Rule\MaxRule;
 use Ares\Rule\MinLengthRule;
@@ -39,8 +42,11 @@ class RuleFactory
         AllowedRule::ID   => ['className' => AllowedRule::class,   'reserved' => false],
         BlankableRule::ID => ['className' => BlankableRule::class, 'reserved' => true],
         DateTimeRule::ID  => ['className' => DateTimeRule::class,  'reserved' => false],
+        DirectoryRule::ID => ['className' => DirectoryRule::class, 'reserved' => false],
         EmailRule::ID     => ['className' => EmailRule::class,     'reserved' => false],
+        FileRule::ID      => ['className' => FileRule::class,      'reserved' => false],
         ForbiddenRule::ID => ['className' => ForbiddenRule::class, 'reserved' => false],
+        LengthRule::ID    => ['className' => LengthRule::class,    'reserved' => false],
         MaxLengthRule::ID => ['className' => MaxLengthRule::class, 'reserved' => false],
         MaxRule::ID       => ['className' => MaxRule::class,       'reserved' => false],
         MinLengthRule::ID => ['className' => MinLengthRule::class, 'reserved' => false],