+++ tests/unit-tests/Ares/Rule/ForbiddenRuleTest.php
@@ -29,7 +29,40 @@ use PHPUnit\Framework\TestCase;
 class ForbiddenRuleTest extends TestCase
 {
     /**
-     * @covers ::validate
+     * @testWith ["Ares\\Rule\\RuleInterface"]
+     *           ["Ares\\Rule\\AbstractRule"]
+     *
+     * @param string $fqcn Fully-qualified class name of the interface or class.
+     * @return void
+     */
+    public function testInstanceOf(string $fqcn): void
+    {
+        $forbiddenRule = new ForbiddenRule();
+
+        $this->assertInstanceOf($fqcn, $forbiddenRule);
+    }
+
+    /**
+     * @covers ::getSupportedTypes
+     *
+     * @testWith ["boolean"]
+     *           ["float"]
+     *           ["integer"]
+     *           ["numeric"]
+     *           ["string"]
+     *
+     * @param string $type Supported type.
+     * @return void
+     */
+    public function testGetSupportedTypes(string $type): void
+    {
+        $forbiddenRule = new ForbiddenRule();
+
+        $this->assertContains($type, $forbiddenRule->getSupportedTypes());
+    }
+
+    /**
+     * @covers ::performValidation
      *
      * @testWith [1]
      *           [17.2]
@@ -50,11 +83,11 @@ class ForbiddenRuleTest extends TestCase
         $this->expectException(InvalidValidationRuleArgsException::class);
         $this->expectExceptionMessage('Invalid args: ' . json_encode($args));
 
-        $forbiddenRule->validate($args, $data, $context);
+        $forbiddenRule->performValidation($args, $data, $context);
     }
 
     /**
-     * @covers ::validate
+     * @covers ::performValidation
      *
      * @dataProvider getValidateSamples
      *
@@ -88,7 +121,7 @@ class ForbiddenRuleTest extends TestCase
 
         $forbiddenRule = new ForbiddenRule();
 
-        $this->assertSame($expectedRetVal, $forbiddenRule->validate($args, $data, $context));
+        $this->assertSame($expectedRetVal, $forbiddenRule->performValidation($args, $data, $context));
     }
 
     /**
@@ -111,7 +144,7 @@ class ForbiddenRuleTest extends TestCase
     }
 
     /**
-     * @covers ::validate
+     * @covers ::performValidation
      *
      * @return void
      */
@@ -139,7 +172,7 @@ class ForbiddenRuleTest extends TestCase
 
         $forbiddenRule = new ForbiddenRule();
 
-        $this->assertFalse($forbiddenRule->validate($args, $data, $context));
+        $this->assertFalse($forbiddenRule->performValidation($args, $data, $context));
     }
 }