+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- adds ```file``` validation rule to check for the existence of a file
+- adds ```directory``` validation rule to check for the existence of a directory
+- adds ```Ares\Rule\AbstractRule``` as validation rule base class
+- improves validation schema parsing to detect inapplicable validation rule usage
+- adds ```tuple``` type to enable validation of fixed length heterogeneous array structures
+- adds ```numeric``` type to enable validation of numeric values (integer or float)
+- adds ```length``` validation rule to check the exact length of a string
 
 ## [v1.0.0] 2019-04-14
 ### Added